📦 feat: created multiple components
