+++ src/components/Main/MainChildren/MovieListStructure/MovieStructure/MovieStructure.jsx
@@ -1,0 +1,18 @@
+import React from 'react';
+
+const MovieStructure = ({ movie }) => {
+  return (
+    <li>
+      <img src={movie.Poster} alt={`${movie.Title} poster`} />
+      <h3>{movie.Title}</h3>
+      <div>
+        <p>
+          <span>🗓</span>
+          <span>{movie.Year}</span>
+        </p>
+      </div>
+    </li>
+  );
+};
+
+export default MovieStructure;
