+++ src/components/Main/MainChildren/MovieListStructure/MovieStructure/WatchedList/WatchedMovieStructure/WatchedMovieStructure.jsx
@@ -1,0 +1,26 @@
+import React from 'react';
+
+const WatchedMovieStructure = ({ movie }) => {
+  return (
+    <li key={movie.imdbID}>
+      <img src={movie.Poster} alt={`${movie.Title} poster`} />
+      <h3>{movie.Title}</h3>
+      <div>
+        <p>
+          <span>⭐️</span>
+          <span>{movie.imdbRating}</span>
+        </p>
+        <p>
+          <span>🌟</span>
+          <span>{movie.userRating}</span>
+        </p>
+        <p>
+          <span>⏳</span>
+          <span>{movie.runtime} min</span>
+        </p>
+      </div>
+    </li>
+  );
+};
+
+export default WatchedMovieStructure;
