+++ src/components/Nav/NavChildrens/NumResult.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+function NumResult() {
+  return (
+    <p className='num-results'>
+      Found <strong>X</strong> results
+    </p>
+  );
+}
+
+export default NumResult;
