+++ src/components/Nav/NavChildrens/SearchInput.jsx
@@ -1,0 +1,16 @@
+import React, { useState } from 'react';
+
+function SearchInput() {
+  const [query, setQuery] = useState('');
+  return (
+    <input
+      className='search'
+      type='text'
+      placeholder='Search movies...'
+      value={query}
+      onChange={(e) => setQuery(e.target.value)}
+    />
+  );
+}
+
+export default SearchInput;
